% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Calculation of Network or Graph Statistics or other Attributes Specified in a Formula}
\usage{
count(formula)
}
\arguments{
\item{formula}{A formula having as its LHS a \code{static.sign} or \code{dynamic.sign} object to be summarized using a formula.}
}
\value{
A vector of statistics specified in the RHS of the formula.
}
\description{
This function computes summaries of the object on the left-hand side (LHS) of the formula that are specified by its right-hand side (RHS). If a network is given as the LHS and \link{sergm.terms} is on the RHS, it computes the sufficient statistics associated with those terms.
}
\seealso{
\link{signNetwork}, \link{sergm.terms}
}
