% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_sampler.R
\name{eval_loglik}
\alias{eval_loglik}
\title{Evaluate Log-Likelihood via Path Sampling}
\usage{
eval_loglik(object)
}
\arguments{
\item{object}{A fitted signed ergm model object.}
}
\value{
A logLik object containing the evaluated log-likelihood, degrees of freedom, and number of observations.
}
\description{
This function evaluates the log-likelihood of a fitted signed ergm model using path sampling.
}
