% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_sergm.R
\name{gof_sergm}
\alias{gof_sergm}
\title{Conduct Goodness-of-Fit Diagnostics on a Signed Exponential Family Random Graph Model}
\usage{
gof_sergm(model, nsim = 200, seed = NULL)
}
\arguments{
\item{model}{A fitted SERGM model.}

\item{nsim}{An integer representing the number of simulated networks to generate. Defaults to 200.}

\item{seed}{Seed value (interger) for the random number generator. See \link{set.seed}}
}
\value{
Plots 6 diagnostics for the goodness-of-fit of signed exponential family random graph models.
}
\description{
This function calculates the goodness-of-fit (GoF) for a fitted SERGM model by simulating new networks using the same model and comparing the statistics of the original network with those of the simulated networks. The statistics used in the comparison are the positive and negative degree distributions, edge-wise shared enemies distributions for positive and negative edges, and edge-wise shared friend distributions for positive and negative edges.
}
\seealso{
\link{sergm}, \link{gof_tsergm}
}
