% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_tsergm.R
\name{gof_tsergm}
\alias{gof_tsergm}
\title{Conduct Goodness-of-Fit Diagnostics on a Temporal Signed Exponential Family Random Graph Model}
\usage{
gof_tsergm(model, nsim = 200)
}
\arguments{
\item{model}{A fitted TSERGM model.}

\item{nsim}{An integer representing the number of simulated networks to generate. Defaults to 200.}
}
\value{
Plots 6 diagnostics for the goodness-of-fit of temporal signed exponential family random graph models.
}
\description{
This function calculates the goodness-of-fit (GoF) for a fitted TSERGM model by simulating new networks using the same model and comparing the average observed statistics of the original network over the timepoints with those of the simulated networks. The statistics used in the comparison are the positive and negative degree distributions, edge-wise shared enemies distributions for positive and negative edges, and edge-wise shared friend distributions for positive and negative edges.
}
\seealso{
\link{tsergm}, \link{gof_sergm}
}
