% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{InitErgmConstraint.fixL}
\alias{InitErgmConstraint.fixL}
\title{Logical layer constraint}
\usage{
# fixL(Ls)
}
\description{
This layer-aware constraint limits the sample space to those networks for which the specified logical layers are unchanged
}
\concept{directed}
\concept{layer-aware}
\concept{undirected}
