% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.sign.R
\name{InitErgmTerm.delnodematch}
\alias{InitErgmTerm.delnodematch}
\title{Delayed node matching on attribute (lag-1)}
\usage{
# binary: delnodematch(attr)
}
\arguments{
\item{attr}{character attribute name to copy from the previous network into the current.}
}
\description{
Create a nodematch term where node attributes come from the previous network's
node attribute `attr`. The previous network used is the one indexed by the current `GroupID`
(equivalent to `lag=1` previously). This constructs a temporary node attribute `delnodecov_<attr>`
on the current network (copied from the previous net) and calls `nodematch` on that attribute.
}
\concept{delayed}
