% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.sign.R
\name{InitErgmTerm.delrecip}
\alias{InitErgmTerm.delrecip}
\title{Delayed reciprocity}
\usage{
# binary: delrecip(base)
}
\arguments{
\item{base}{character or numeric name/identifier of the layer to examine in the current network.}
}
\description{
For the current network layer `base`, this term equals 1 for each directed edge
i->j currently present where the reverse edge j->i was present in the previous network's same
layer. The previous network used is the one indexed by the current `GroupID` (i.e., the behaviour
is the same as the prior `lag=1` implementation). The term is provided as an edgecov (1/0).
}
\concept{delayed}
