% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.sign.R
\name{InitErgmTerm.Neg}
\alias{InitErgmTerm.Neg}
\title{Evaluation of negative edges}
\usage{
# binary: Neg(formula)
}
\arguments{
\item{formula}{a one-sided [ergm()]-style formula with the terms to be evaluated}
}
\description{
Evaluates the terms in `formula` of the negative edges and sums the results elementwise.
}
\concept{operator}
