% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.sign.R
\name{InitErgmTerm.Pos}
\alias{InitErgmTerm.Pos}
\title{Evaluation of positive edges}
\usage{
# binary: Pos(formula)
}
\arguments{
\item{formula}{a one-sided [ergm()]-style formula with the terms to be evaluated}
}
\description{
Evaluates the terms in `formula` of the positive edges and sums the results elementwise.
}
\concept{operator}
