% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mple_sign.R
\name{mple_sign}
\alias{mple_sign}
\title{Fit an ERGM with MPLE using a logistic regression model}
\usage{
mple_sign(
  formula,
  control = control.ergm(MPLE.covariance.method = "Godambe"),
  ...
)
}
\arguments{
\item{formula}{An ERGM formula, with the left-hand side being the network}

\item{control}{A list of control parameters for the \code{\link{ergmMPLE}} function. Per default the method to estimate the convariance method is set to Godambe.}

\item{...}{Additional arguments passed to \code{\link{ergmMPLE}}}
}
\value{
A fitted logistic regression model
}
\description{
Return a fitted logistic regression model to calculate the MPLE of an ERGM.
}
\details{
The MPLE for an ERGM is calculated by first finding the matrices of
positive and negative change statistics. The resulting change matrices are
then used to calculate the MPLE using a logistic regression model.
}
\examples{
data(tribes)
mple_sign(tribes ~ Pos(~edges) + Neg(~edges))

}
\seealso{
\code{\link{ergmMPLE}}
}
