% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.dynamic.sign}
\alias{plot.static.sign}
\title{Plot Signed Network for Signed Networks}
\usage{
plot(
  net,
  time = c(1:length(net)),
  color_pos = "green3",
  color_neg = "red3",
  vertex.col = 2,
  vertex.legend = F,
  vertex.legend.pos = "topleft",
  main = paste("Time ", c(1:length(net))),
  vertex.legend.size = 0.65,
  ...
)

\method{plot}{dynamic.sign}(
  net,
  time = c(1:length(net)),
  color_pos = "green3",
  color_neg = "red3",
  vertex.col = 2,
  vertex.legend = F,
  vertex.legend.pos = "topleft",
  main = paste("Time ", c(1:length(net))),
  vertex.legend.size = 0.65,
  ...
)

\method{plot}{static.sign}(
  net,
  color_pos = "green3",
  color_neg = "red3",
  vertex.legend = F,
  vertex.col = 2,
  vertex.legend.pos = "topleft",
  legend.size = 0.65,
  ...
)
}
\arguments{
\item{net}{An object of class \code{static.sign} or \code{dynamic.sign} created with \code{\link{signNetwork}}.}

\item{time}{A vector of integers indicating which timepoints should be plotted.}

\item{color_pos}{Edge color for positive edges (1). Default is set to green.}

\item{color_neg}{Edge color for negative edges (-1). Default is set to red.}

\item{vertex.col}{Choose a vertex attribute as a color for the vertices.}

\item{vertex.legend}{Display vertex legend, if \code{vertex.col} is a vertex attribute.}

\item{vertex.legend.pos}{Specifying legend’s position using one of the following options: ”bottomright”, ”bottom”, ”bottomleft”, ”left”, ”topleft”, ”top”, ”topright”, ”right” or ”center”.}

\item{main}{A list of strings indicating the plot titles. If not specified, the plots will be named ”Time 1” to ”Time T”. If no titles are desired, the input should be set to NULL.}

\item{vertex.legend.size}{Size of the legend, which takes a character expansion factor relative to the current \code{par("cex")}.}

\item{...}{Additional arguments to plot \code{\link{network::plot.network}}}
}
\value{
One or multiple two-dimensional plots.
}
\description{
The function \code{plot.static.sign} or \code{plot.dynamic.sign} produces one or multiple simple two-dimensional plot of a dynamic signed network. In addition to the arguments of the \code{\link{network::plot.network}} function, some arguments for the context of signed networks have been added.
}
\seealso{
\link{signnet}
}
