% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{randomtoggleFixL-ergmProposal}
\alias{randomtoggleFixL-ergmProposal}
\alias{InitErgmProposal.randomtoggleFixL}
\title{Propose a randomly selected dyad to toggle, respecting the layer constraint}
\description{
Propose a randomly selected dyad to toggle
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="randomtoggleFixL"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="randomtoggleFixL"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="randomtoggleFixL"))}}
}
\keyword{internal}
