% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rebels}
\alias{rebels}
\title{Conflict Events in Syrian Civil War}
\format{
An undirected \code{dynamic.sign} object with no loops and three timepoints.
}
\description{
A dynamic network of combat events in the Syrian civil war between 2017 and
2019. The dataset is taken from \insertCite{fritz2023all;textual}{ergm.sign}.
The raw data for this example Armed Conflict Location & Event Data Project \insertCite{raleigh2010introducing;textual}{ergm.sign}.
}
\examples{

\donttest{
data(rebels)
}
}
\references{
\insertRef{fritz2023all}{ergm.sign} \insertRef{raleigh2010introducing}{ergm.sign}
}
