% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signnet.R
\name{signnet}
\alias{signnet}
\title{Create Signed Network Object}
\usage{
signnet(
  mat,
  directed = F,
  loops = F,
  matrix.type,
  cov = NULL,
  names = NULL,
  ...
)
}
\arguments{
\item{mat}{(List of) Adjacency matrix or edgelist. The adjacency matrix must only consist of 1,0 or -1. The edgelist must consist of 3 columns "From", "To" and "Sign" (1 or -1). For a dynamic network the required input is a list of adjacecny matrices or edgelists.}

\item{directed}{logical; should edges be interpreted as directed?}

\item{loops}{logical; should loops be allowed?}

\item{matrix.type}{Either "adjacency" or "edgelist" indicating what kind of format the input has.}

\item{cov}{Add vertex attributes to the network. The input for this should be a dataframe where the first column contains the names of the vertices and the other columns of the dataframe should represent the individual vertex attributes that will be added to the network, however it is not necessary to include all the vertices of the network in the dataframe.}

\item{names}{Specify vertex names. If no list is provided, the column names of the adjacency matrix will be used as names for the vertices.}
}
\value{
A signed network of the class \code{static.sign} or \code{dynamic.sign}.
}
\description{
Turn an adjacency matrix or an edgelist into a static or dynamic signed network
}
