% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signNetworks.R
\name{signNetworks}
\alias{signNetworks}
\title{A signed multinetwork network representation.}
\usage{
signNetworks(...)
}
\description{
A function for specifying the LHS of a signed multi-network (a.k.a. multilevel) ERGM. Typically used in conjunction with the \code{N()} term operator.
}
\details{
@param ... Several signed networks or a list of signed networks.

 @return A network object comprising the provided networks, with multinetwork metadata.

 @export
}
