% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tsergm.R
\name{sim_tsergm}
\alias{sim_tsergm}
\title{Draw from the Distribution of a Temporal Signed Exponential Family Random Graph Model}
\usage{
sim_tsergm(object, nsim = 1, seed = NULL, coef = NULL, ...)
}
\arguments{
\item{object}{Either a formula or a fitted tsergm. The formula should be of the form y ~ <model terms>, where y is a signed network of the class \code{dynamic.sign}.}

\item{nsim}{Number of networks to be randomly drawn from the given distribution on the set of all networks, returned by the Metropolis-Hastings algorithm.}

\item{seed}{Seed value (integer) for the random number generator. See \link{set.seed}.}

\item{coef}{Vector of parameter values for the model from which the sample is to be drawn.}

\item{...}{Further arguments passed to or used by methods.}
}
\value{
A dynamic signed network or a list of dynamic signed networks (if nsim > 1) of class \code{dynamic.sign}.
}
\description{
\code{sim_tsergm} is used to draw from temporal signed exponential family random network models. See \link{tsergm} for more information on these models. The method for tsergm objects inherits the model, the coefficients, the response attribute, the reference, the constraints, and most simulation parameters from the model fit, unless overridden by passing them explicitly. Unless overridden, the simulation is initialized with either a random draw from near the fitted model saved by \code{tsergm()}.
}
\seealso{
\link{signnet}, \link{tsergm}, \link{sergm.terms}, \link{sim_tsergm}
}
