% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sponsor}
\alias{sponsor}
\title{Common Sponsor Data for Syrian Civil War Factions}
\format{
A matrix with 68 rows and 68 columns.
}
\description{
A data frame containing binary indicators for whether each faction in the Syrian civil war is sponsored by a common external actor.
}
\examples{
data(sponsor)
}
\references{
\insertRef{fritz2023all}{ergm.sign}
}
