% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.dynamic.sign}
\alias{summary.static.sign}
\title{Network Attributes for Signed Networks}
\usage{
summary(net, time = c(1:length(net)))

\method{summary}{dynamic.sign}(net, time = c(1:length(net)), names = NULL)

\method{summary}{static.sign}(net)
}
\arguments{
\item{net}{A signed network object of class \code{static.sign} or \code{dynamic.sign}.}

\item{time}{A list of integers indicating what timepoints should be summarised.}
}
\value{
Matrix with network attributes.
}
\description{
Print descriptive statistics of a signed network.
}
\seealso{
\link{signnet}
}
