% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.static.sign}
\alias{summary.static.sign}
\alias{summary.dynamic.sign}
\title{Network Attributes for Signed Networks}
\usage{
\method{summary}{static.sign}(net)

\method{summary}{dynamic.sign}(net, time = NULL, names = NULL)
}
\arguments{
\item{net}{A signed network object of class \code{static.sign}.}

\item{time}{A vector of integers indicating which timepoints should be summarized.}

\item{names}{A character vector of names for the timepoints. If \code{NULL}, timepoints are numbered.}
}
\value{
A data frame or matrix with network attributes.
}
\description{
Print descriptive statistics of a signed network.
}
\section{Static signed networks}{

\code{summary.static.sign()} summarizes a single (static) signed network.
}

\examples{
data("tribes")
summary.static.sign(tribes)

}
\seealso{
\link{signNetwork}
}
