% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tribes}
\alias{tribes}
\title{Read Highland Tribes}
\format{
An undirected \code{static.sign} object with no loops.
}
\source{
\url{http://vlado.fmf.uni-lj.si/pub/networks/data/UciNet/UciData.htm#gama},
with corrections from Read (1954).
}
\description{
A network of political alliances and enmities among the 16 Gahuku-Gama
sub-tribes of Eastern Central Highlands of New Guinea, documented by Read
(1954).
}
\details{
This network shows 3 clusters.
}
\examples{

\donttest{
data(tribes)
}

}
\references{
Taken from UCINET IV, which cites the following: Hage P. and
Harary F. (1983). Structural models in anthropology. Cambridge: Cambridge
University Press. (See p 56-60).  Read K. (1954). Cultures of the central
highlands, New Guinea. Southwestern Journal of Anthropology, 10, 1-43.
}
\keyword{cluster}
\keyword{graphs}
\keyword{multivariate}
