% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnLayer.R
\name{UnLayer}
\alias{UnLayer}
\title{Multilayer network to single layer network.}
\usage{
UnLayer(
  net,
  color_pos = "green3",
  color_neg = "red3",
  neg.lyt = c("solid", "dashed", "dotted")
)
}
\arguments{
\item{net}{A signed network object of class \code{static.sign} or \code{dynamic.sign}.}

\item{color_pos}{Color for positive edges. Default is 'green3'.}

\item{color_neg}{Color for negative edges. Default is 'red3'.}
}
\value{
Single layer network object.
}
\description{
Turn a multilayer network object into a single layer network object.
}
\examples{
data("tribes")
tribes_sgl <- UnLayer(tribes)

}
\seealso{
\link{signNetwork}
}
