% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.sign.R
\docType{package}
\name{ergm.sign}
\alias{ergm.sign}
\title{ergm.sign: A Package for Exponential Random Graph Models for Signed Networks}
\description{
The ergm.sign package implements the tools to simulate and estimate Signed Exponential Random Graph Models and Temporal Signed Exponential Random Graph Models.
}
\section{Authors}{

Marc Schalberger
}

