% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.sign-package.R
\name{ergm.sign}
\alias{ergm.sign}
\title{ergm.sign: A Package for Exponential Random Graph Models for Signed Networks}
\description{
The ergm.sign package implements tools to simulate and estimate Signed Exponential Random Graph Models
and Temporal Signed Exponential Random Graph Models.
}
\author{
Marc Schalberger
}
