% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.static.sign}
\alias{plot.static.sign}
\title{Visualization for Signed Networks}
\usage{
\method{plot}{static.sign}(
  net,
  col_pos = "green3",
  col_neg = "red3",
  neg.lyt = "dotted",
  inv_weights = T,
  ...
)
}
\arguments{
\item{net}{A signed network object of class \code{static.sign} or \code{dynamic.sign}.}

\item{col_pos}{Color for positive edges. Default is 'green3'.}

\item{col_neg}{Color for negative edges. Default is 'red3'.}

\item{neg.lyt}{Line type for negative edges. Default is "dotted". Other options are "solid" and "dashed".}

\item{inv_weights}{Logical. If TRUE, edge weights are inverted (1/weights) to ensure that the endpoints of positive edges are closer togehter than negative. Default is TRUE.}

\item{...}{Additional arguments passed to the plot function.}

\item{titles}{A character vector of names for the timepoints.}

\item{time}{A list of integers indicating what timepoints should be visualized.}
}
\value{
A plot of the signed network.
}
\description{
This function visualizes a signed network.
}
\examples{
data("tribes")
plot.static.sign(tribes, col_pos = "green", col_neg = "red")

}
\seealso{
\link{UnLayer}
}
