% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rebels}
\alias{rebels}
\title{Conflict Events in Syrian Civil War}
\format{
An undirected \code{dynamic.sign} object with no loops and eight timepoints.
}
\description{
A dynamic network of combat events in the Syrian civil war between 2017 and
2025. The raw data comes from the Armed Conflict Location & Event Data Project \insertCite{raleigh2010introducing;textual}{ergm.sign}.
}
\examples{
data(rebels)
}
\references{
\insertRef{fritz2023all}{ergm.sign}, \insertRef{raleigh2010introducing}{ergm.sign}
}
