% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dynamic.sign}
\alias{summary.dynamic.sign}
\title{Network Attributes for Signed Networks}
\usage{
\method{summary}{dynamic.sign}(net, time = NULL, names = NULL)
}
\arguments{
\item{net}{A signed network object of class \code{static.sign} or \code{dynamic.sign}.}

\item{time}{A list of integers indicating what timepoints should be summarised.}

\item{names}{A character vector of names for the timepoints. If NULL, the time points will be numbered.}
}
\value{
Matrix with network attributes.
}
\description{
Print descriptive statistics of a signed network.
}
\examples{
data("tribes")
summary(tribes)

}
\seealso{
\link{signNetwork}
}
