% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.static.sign}
\alias{summary.static.sign}
\alias{summary.dynamic.sign}
\title{Network Attributes for Signed Networks}
\usage{
\method{summary}{static.sign}(object, ...)

\method{summary}{dynamic.sign}(object, time = NULL, names = NULL, ...)
}
\arguments{
\item{object}{A signed network object of class \code{dynamic.sign}.}

\item{...}{Additional arguments.}

\item{time}{Integer vector of timepoints to summarize. Defaults to all.}

\item{names}{Character vector of names for timepoints. If NULL, uses "Time 1", "Time 2", etc.}
}
\value{
A data frame or matrix with network attributes.
}
\description{
Print descriptive statistics of a signed network.
}
\section{Static signed networks}{

\code{summary.static.sign()} summarizes a single (static) signed network.
}

\examples{
data("tribes")
summary(tribes)

}
\seealso{
\link{network.sign}, \link{UnLayer}
}
