% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tribes}
\alias{tribes}
\title{Read Highland Tribes}
\format{
An undirected \code{static.sign} object with no loops.
}
\description{
A static network of political alliances and enmities among the 16 Gahuku-Gama
sub-tribes of Eastern Central Highlands of New Guinea, documented by \insertCite{read1954cultures;textual}{ergm.sign}.
}
\examples{
\donttest{
data(tribes)
}
}
\references{
Taken from UCINET IV, which cites the following: \insertRef{hage1983structural}{ergm.sign}, \insertRef{read1954cultures}{ergm.sign}
}
